+++ src/main/java/ai/lucidtech/las/sdk/DeleteDocumentsOptions.java
@@ -7,7 +7,6 @@ import java.util.List;
 
 
 public class DeleteDocumentsOptions extends DeleteResourcesOptions<DeleteDocumentsOptions> {
-    private String[] batchId;
     private String[] datasetId;
     private String[] consentId;
 
@@ -21,13 +20,7 @@ public class DeleteDocumentsOptions extends DeleteResourcesOptions<DeleteDocumen
         return this;
     }
 
-    public DeleteDocumentsOptions setBatchId(String[] batchId) {
-        this.batchId = batchId;
-        return this;
-    }
-
     public List<NameValuePair> addOptions(List<NameValuePair> parameters) {
-        this.addOption(parameters, "batchId", this.batchId);
         this.addOption(parameters, "datasetId", this.datasetId);
         this.addOption(parameters, "consentId", this.consentId);
         return super.addOptions(parameters);