+++ src/main/java/ai/lucidtech/las/sdk/ListBatchesOptions.java
@@ -1,4 +1,0 @@
-package ai.lucidtech.las.sdk;
-
-
-public class ListBatchesOptions extends ListResourcesOptions<ListBatchesOptions> {}