+++ src/main/java/ai/lucidtech/las/sdk/UpdateBatchOptions.java
@@ -1,3 +1,0 @@
-package ai.lucidtech.las.sdk;
-
-public class UpdateBatchOptions extends NameAndDescriptionOptions<UpdateBatchOptions> {}