+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog 
 
+## Version 3.0.0 - 2021-10-14
+
+- Removed all support for batches. Use datasets instead.
+- Added UpdateDocumentOptions to updateDocument
+- Added datasetId, groundTruth and retentionInDays to UpdateDocumentOptions
+
 ## Version 2.0.0 - 2021-08-31
 
 - Bugfixes for List and Delete Options